Remove the padding from the end of the text.
The algorithm as given in the paper handles the fact that we want the
score for the `pattern.length % w`'th row of the last block instead of
the w'th row by "padding" the pattern and text with wildcard characters
to make it an even multiple of the word size.

IMO this is a bit confusing and can be handled more directly by
adjusting the initial score and last row bit-masks for the bottom block
to reflect that not all rows in that block are used.