Fix bug when maxErrors exceeds pattern length by word size.
In `findMatchEnds` the max block index `bMax` was calculated based on
the pattern length but the initial value of the bottom active block `y`
was determined based on the `maxErrors` arg. Therefore it was possible
for `y` to exceed `bMax` if `maxErrors > pattern.length` and the
algorithm failed to produce any matches.

Fix this by clamping "maxErrors".