Optimize array accesses and map lookups
Replace generic arrays with typed arrays for bitmasks and `Map` for an
associative map between char values and bitmasks.

Typed arrays are automatically initialized to zero on creation which
avoids the need for a `fill` step and an associative map is, I believe,
cheaper than using a holey-array for this purpose.