Handle patterns of unequal lengths correctly in `findMatchRegions`
When patterns passed to `multiSearch` have different lengths, the
filtering step must effectively "pad" short patterns with wildcard chars
to make all the patterns equal length. This must be done both for chars
that appear in (any of) the patterns and for chars which appear in the
text but not in any of the patterns.