Fix search when pattern is exact multiple of word size
Fix incorrect results from search when all of the following conditions
are met:

 1. The pattern length is an exact multiple of the word size (32)

 2. The pattern matches the text with errors

 3. The `maxErrors` param is large enough to allow a match to be found,
    but smaller than the pattern length by at least word-size chars.

There was an error in the calculation of the `maxBlockScore` variable which
should be equal to the number of characters that the pattern extends into a
block. When the pattern exactly fills the last used block it should be equal to
the word size, but instead was calculated as 0.